requirements.txt
asn1crypto==0.24.0
azure-common==1.1.16
azure-nspkg==2.0.0
azure-storage==0.36.0
boto3==1.8.9
botocore==1.11.9
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
docutils==0.14
et-xmlfile==1.0.1
Flask==1.0.2
future==0.17.0
gunicorn==19.9.0
idna==2.7
ijson==2.3
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
openpyxl==2.5.9
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodomex==3.6.6
pyee==5.0.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.6
requests==2.20.0
s3transfer==0.1.13
six==1.11.0
slackclient==1.3.0
slackeventsapi==2.0.0
snowflake-connector-python==1.6.11
urllib3==1.23
websocket-client==0.53.0
Werkzeug==0.14.1
